excludeDelFilesOption

+++ segment.c
@@ -65,7 +65,8 @@ uint8_t *SegmentSequence(uint8_t *fName, Parameters *P)
 
   fclose(Reader);
   fclose(Writter);
-  unlink(fName);
+  
+  if(P->del) unlink(fName);
 
   if(P->verbose == 1)
     {
